refactor(transactionRoutes): moved validation to a middleware
